// src/types/tableOfContents.enum.ts
export enum CHAPTER {
    CHAPTER1 = 'CHAPTER1',
    CHAPTER2 = 'CHAPTER2',
    CHAPTER3 = 'CHAPTER3',
}
export enum LESSON {
    LESSON1 = 'LESSON1',
    LESSON2 = 'LESSON2',
    LESSON3 = 'LESSON3',
    LESSON4 = 'LESSON4',
    LESSON5 = 'LESSON5',
    LESSON6 = 'LESSON6',
    LESSON7 = 'LESSON7',
    LESSON8 = 'LESSON8',
}